// src/app/models/order.model.ts
export interface Order {
    id: string;
    idUser: string;
    status: string;
    priceTotal: string;
    idProduct1: string;
    quantity1: string;
    priceUni1: string;
    idProduct2: string;
    quantity2: string;
    priceUni2: string;
    idProduct3: string;
    quantity3: string;
    priceUni3: string;
    idProduct4: string;
    quantity4: string;
    priceUni4: string;
    idProduct5: string;
    quantity5: string;
    priceUni5: string;
    userOrder: string;
}